post view, deleting

+++ aiogram_bot/handlers/z_delete.py
@@ -7,12 +7,12 @@ from bot import dp
 
 
 @dp.message_handler(content_types=ContentType.ANY, state="*")
-async def delete(message: Message):
+async def delete(message: Message, error_message: bool = True):
     try:
         await message.delete()
     except MessageCantBeDeleted:
         pass
-
-    msg = await message.answer("Читаем инструкцию и не спамим!")
-    await sleep(2)
-    await msg.delete()
+    if error_message:
+        msg = await message.answer("Не надо писать лишнее, я по пустякам не отвечаю.")
+        await sleep(2)
+        await msg.delete()
